refactor: remove duplicate `getEntitiesQuery` and `subgraphFieldsFragment`
