+++ apps/hash-frontend/src/graphql/queries/knowledge/entity.queries.ts
@@ -1,6 +1,5 @@
 import { gql } from "@apollo/client";
-
-import { subgraphFieldsFragment } from "../subgraph";
+import { subgraphFieldsFragment } from "@local/hash-graphql-shared/queries/subgraph";
 
 export const createEntityMutation = gql`
   mutation createEntity(
@@ -17,37 +16,6 @@ export const createEntityMutation = gql`
       linkData: $linkData
     )
   }
-`;
-
-export const getEntityQuery = gql`
-  query getEntity(
-    $entityId: EntityId!
-    $entityVersion: String
-    $constrainsValuesOn: OutgoingEdgeResolveDepthInput!
-    $constrainsPropertiesOn: OutgoingEdgeResolveDepthInput!
-    $constrainsLinksOn: OutgoingEdgeResolveDepthInput!
-    $constrainsLinkDestinationsOn: OutgoingEdgeResolveDepthInput!
-    $inheritsFrom: OutgoingEdgeResolveDepthInput!
-    $isOfType: OutgoingEdgeResolveDepthInput!
-    $hasLeftEntity: EdgeResolveDepthsInput!
-    $hasRightEntity: EdgeResolveDepthsInput!
-  ) {
-    getEntity(
-      entityId: $entityId
-      entityVersion: $entityVersion
-      constrainsValuesOn: $constrainsValuesOn
-      constrainsPropertiesOn: $constrainsPropertiesOn
-      constrainsLinksOn: $constrainsLinksOn
-      constrainsLinkDestinationsOn: $constrainsLinkDestinationsOn
-      inheritsFrom: $inheritsFrom
-      isOfType: $isOfType
-      hasLeftEntity: $hasLeftEntity
-      hasRightEntity: $hasRightEntity
-    ) {
-      ...SubgraphFields
-    }
-  }
-  ${subgraphFieldsFragment}
 `;
 
 export const queryEntitiesQuery = gql`
