+++ apps/hash-frontend/src/graphql/queries/ontology/entity-type.queries.ts
@@ -1,6 +1,5 @@
 import { gql } from "@apollo/client";
-
-import { subgraphFieldsFragment } from "../subgraph";
+import { subgraphFieldsFragment } from "@local/hash-graphql-shared/queries/subgraph";
 
 export const getEntityTypeQuery = gql`
   query getEntityType(
