+++ libs/@local/hash-graphql-shared/src/queries/entity.queries.ts
@@ -1,41 +1,7 @@
 import { gql } from "@apollo/client";
 
-export const subgraphFieldsFragment = gql`
-  fragment SubgraphFields on Subgraph {
-    roots
-    vertices
-    edges
-    depths {
-      constrainsLinkDestinationsOn {
-        outgoing
-      }
-      constrainsLinksOn {
-        outgoing
-      }
-      constrainsValuesOn {
-        outgoing
-      }
-      constrainsPropertiesOn {
-        outgoing
-      }
-      inheritsFrom {
-        outgoing
-      }
-      isOfType {
-        outgoing
-      }
-      hasLeftEntity {
-        incoming
-        outgoing
-      }
-      hasRightEntity {
-        incoming
-        outgoing
-      }
-    }
-    temporalAxes
-  }
-`;
+import { subgraphFieldsFragment } from "./subgraph";
+
 export const getEntityQuery = gql`
   query getEntity(
     $entityId: EntityId!
