+++ libs/@local/hash-graphql-shared/src/queries/subgraph.ts
